Add files via upload
Location: 
/home/disinfo_study_2310/vanessachan/whisper-scripts

Architecture: 
This code processes mpg2 files from /home/disinfo_study_2310/vanessachan/mpg-transfer on ffmpeg downloaded onto the Itkin GPU cluster and transfer them into /home/disinfo_study_2310/vanessachan/rosenthal-landing. mpg-transfer has an NFS mount to the storage of Rosenthal tapes on the Hoffman GPU cluster, and rosenthal-landing has an NFS mount to an empty file in the Hoffman GPU cluster. The code processes the files from Hoffman through ffmpeg locally on the Itkin cluster and stores it back on Hoffman. disinfo_study_2310 should have read access to mpg-transfer and read and write access to rosenthal-landing. 

The requirements are that each file is processed twice on ffmpeg (two-pass) to ensure quality with a processing bitrate to ensure the final mp4 file is 20% of the file size of the original mpg2 file (with an upper bitrate cap of 10Mbps and a target of 5 Mbps).


To run:
# creates screen on your terminal
screen
# runs code
./mpg_convert.sh

# Click to release screen from your terminal so code can run in background
Ctrl A D